+++ src/config/BridgeConfig.ts
@@ -3,6 +3,7 @@ import { LoggerConfig } from "../logging";
 import { IrcHandlerConfig } from "../bridge/IrcHandler";
 import { RoomConfigConfig } from "../bridge/RoomConfig";
 import { MatrixHandlerConfig } from "../bridge/MatrixHandler";
+import { Rules } from "matrix-appservice-bridge";
 
 export interface BridgeConfig {
     database: {
@@ -25,7 +26,7 @@ export interface BridgeConfig {
         provisioning: {
             enabled: boolean;
             requestTimeoutSeconds: number;
-            ruleFile: string;
+            rules: Rules;
             enableReload: boolean;
             roomLimit?: number;
         };
@@ -58,6 +59,11 @@ export interface BridgeConfig {
             [userIdOrDomain: string]: "admin";
         };
         perRoomConfig?: RoomConfigConfig;
+        RMAUlimit?: number;
+        userActivity?: {
+            minUserActiveDays?: number;
+            inactiveAfterDays?: number;
+        };
     };
     sentry?: {
         enabled: boolean;