+++ src/datastore/NedbDataStore.ts
@@ -14,13 +14,16 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import Bluebird from "bluebird";
 import { IrcRoom } from "../models/IrcRoom";
 import { IrcClientConfig, IrcClientConfigSeralized } from "../models/IrcClientConfig"
 import { getLogger } from "../logging";
 
-import { MatrixRoom, MatrixUser, RemoteUser, RemoteRoom,
-    UserBridgeStore, RoomBridgeStore, RoomBridgeStoreEntry as Entry } from "matrix-appservice-bridge";
+import {
+    MatrixRoom, MatrixUser, RemoteUser, RemoteRoom,
+    UserBridgeStore, UserActivityStore,
+    RoomBridgeStore, RoomBridgeStoreEntry as Entry,
+    UserActivity, UserActivitySet
+} from "matrix-appservice-bridge";
 import { DataStore, RoomOrigin, ChannelMappings, UserFeatures } from "./DataStore";
 import { IrcServer, IrcServerConfig } from "../irc/IrcServer";
 import { StringCrypto } from "./StringCrypto";
@@ -36,6 +39,7 @@ export class NeDBDataStore implements DataStore {
     private cryptoStore?: StringCrypto;
     constructor(
         private userStore: UserBridgeStore,
+        private userActivityStore: UserActivityStore,
         private roomStore: RoomBridgeStore,
         private bridgeDomain: string,
         pkeyPath?: string) {
@@ -209,12 +213,9 @@ export class NeDBDataStore implements DataStore {
      * @return {Promise} A promise which resolves to a list
      * of entries.
      */
-    public getProvisionedMappings(roomId: string): Bluebird<Entry[]> {
-        return Bluebird.cast(this.roomStore.getEntriesByMatrixId(roomId)).filter(
-            (entry: Entry) => {
-                return entry.data && entry.data.origin === 'provision'
-            }
-        );
+    public async getProvisionedMappings(roomId: string): Promise<Entry[]> {
+        const mappings = await this.roomStore.getEntriesByMatrixId(roomId);
+        return mappings.filter(entry => entry.data && entry.data.origin === 'provision');
     }
 
     /**
@@ -588,6 +589,14 @@ export class NeDBDataStore implements DataStore {
         }
         matrixUser.set("features", features);
         await this.userStore.setMatrixUser(matrixUser);
+    }
+
+    public async getUserActivity(): Promise<UserActivitySet> {
+        return this.userActivityStore.getActivitySet();
+    }
+
+    public async storeUserActivity(userId: string, activity: UserActivity) {
+        this.userActivityStore.storeUserActivity(userId, activity);
     }
 
     public async storePass(userId: string, domain: string, pass: string) {